Split PanelJupyterExecutor into separate module
